netifd: Assign interface metric to route metric when route is created
Interface metric needs to be assigned to the route metric parameter at route creation time. Otherwise if the interface metric is different from 0 route_cmp will wrongly conclude the routes are different.
In this case the route will be added/deleted and could end up with the route missing in the kernel depending on the add/delete order.

Signed-off-by: Hans Dedecker <dedeckeh@gmail.com>