netifd: Mark interface as dynamic after it has been added in the interface list
Adding an interface in the interface list in case of a config update will override the node version of the old interface and thus overriding the dynamic interface marker.
In case of config update interface_set_dynamic is done on an invalid interface pointer as the new interface pointer has been freed.

Signed-off-by: Hans Dedecker <dedeckeh@gmail.com>