netifd: Fix lost interface device config for devices having default config
As the PPP main device (either a simple or old style vlan device) is freed when PPP interfaces are reloaded; interface device specific config (eg mtu/macaddr/...) is lost as the interface config is not re-applied on the newly created device.
Patch fixes the issue by re-applying the interface device config after the main device is claimed and is having default config.

Signed-off-by: Hans Dedecker <dedeckeh@gmail.com>