netifd: Fix device ifindex overwrite when processing netlink event messages
When a device with the same name is deleted and created again in the kernel
the ifindex changes.
A race condition will occur when netlink event messages linked to the old device
are processed and will thus overwrite the correct ifindex of the new device.
Further make sure a valid ifindex is in place for both external and internal
devices when setting the state to enabled.

Signed-off-by: Hans Dedecker <dedeckeh@gmail.com>