netifd: Fix possible hotplug race conditions
Don't drop ifup/ifdown events in case an interface event is cached
as it leads to possible race conditions (eg firewall not being
reloaded as ifup is dropped) when multiple interface events are fired
in a short timeframe (eg multiple PPP link flaps).
Always overwrite the cached interface event except for the
interface update event so the hotplug scripts are launched
with the last known status.

Signed-off-by: Hans Dedecker <dedeckeh@gmail.com>