interface-ip: harden eui64 IPv6 prefix address generation
Check if a mac address is actually present when generating an eui64 based
IPv6 address; in case of failure bail out.
At the same time make sure the active mac address is used as input for the
eui64 based IPv6 address and guarantee IPv6 prefix address generation is
based on the actual config by resetting the IPv6 prefix address in the
assignment structure when it gets deleted.

Signed-off-by: Hans Dedecker <dedeckeh@gmail.com>