treewide: fix use of IFNAMSIZ
IFNAMSIZ already includes an extra byte for the null character
terminating the string. Adding an extra byte can cause the interface
name to be truncated, resulting in undefined behaviour.

Drop the extra byte. For the bonding case, the interface name written to
sysfs should be prefixed with + or - so we do need an extra byte.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>