config: network-device is incompatible with jshn
An oversight in openwrts uci-defaults.sh ucidef_set_network_device_*
functions resulted in a hyphen (-) being used insted of a underscore (_)
this causes problems for jshn which will translate the former to the
latter. This causes config_init_board to return null on boards that
set network?device using uci-defaults (or jshn).

This patch corrects the issue by defaulting to network_device with a
fallback to network-device if nothing is returned.

Signed-off-by: Michael 'ASAP' Weinrich