system-linux: inherit tos value in outer ip6 header
In case of ipip6 tunnels, inherit the tos value in the ip6 header
from the inner ip4 header as the default behaviour. Without
this commit, the tos value in the ip6 header is hard coded to 0 which
is wrong.

The ipip6 mode tunnels lack a tos uci option allowing the choice of
either inheriting tos value or setting it as per uci config. I will
push a change to address this separately, but, in the absense of such
a config, the default behaviour should be inherit the tos value from
the inner ip4 header and not to hard code it to 0. This commit fixes
that.

Signed-off-by: Rahul Thakur <rahul.thakur@iopsys.eu>